+++ NEWS.md
@@ -1,3 +1,12 @@
+0.2.0
+=========
+
+- Updated authentication provider to be `com.amazonaws.athena.jdbc.shaded.com.amazonaws.auth.DefaultAWSCredentialsProviderChain` (via @dabdine)
+- Now supports additional DBI/RJDBC methods including: `dbExistsTable()`,
+  `dbListFields()`, `dbListTables()`, `dbReadTable()`
+- More documentation
+- Added code of conduct
+
 0.1.0 
 =========